requirements.txt
pandas
numpy
faiss-cpu
torch
transformers
Pillow
requests
ollama
fastapi
streamlit
